requirements.txt
asgiref==3.5.2
Django==3.2.16
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-markdownx==1.2.1
django-solo==2.0.0
django-tinymce==3.5.0
Markdown==3.4.1
Pillow==9.3.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.6
